+++ lnd_methods/onchain/open_channel.js
@@ -24,9 +24,9 @@ const type = 'default';
   `base_fee_mtokens` is not supported on LND 0.15.5 and below
   `fee_rate` is not supported on LND 0.15.5 and below
 
-  `is_max_funding` is not supported on LND 0.16.3 and below
+  `is_max_funding` is not supported on LND 0.16.4 and below
 
-  `description` is not supported on LND 0.16.3 and below
+  `description` is not supported on LND 0.16.4 and below
 
   {
     [base_fee_mtokens]: <Routing Base Fee Millitokens Charged String>